+++ settings/header_menus.php
@@ -133,5 +133,5 @@ if ($ADMIN->fulltree) {
         $page->add($setting);
     }
 
-    $settings->add($page);
+    $asettings->add($page);
 }