+++ settings/layout_responsive.php
@@ -173,5 +173,5 @@ if ($ADMIN->fulltree) {
     $setting->set_updatedcallback('theme_reset_all_caches');
     $page->add($setting);
 
-    $settings->add($page);
+    $asettings->add($page);
 }