+++ settings/marketing_blocks.php
@@ -96,5 +96,5 @@ if ($ADMIN->fulltree) {
         $page->add($setting);
     }
 
-    $settings->add($page);
+    $asettings->add($page);
 }