+++ settings/navbar_links.php
@@ -80,5 +80,5 @@ if ($ADMIN->fulltree) {
         $page->add($setting);
     }
 
-    $settings->add($page);
+    $asettings->add($page);
 }